Move the looped service into its own package
