+++ Go/Loop/LoopFivonacciService.go
@@ -1,4 +1,4 @@
-package main
+package loop
 
 import (
 	"Fibonacci/common"
